requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

easyocr
matplotlib
numpy
pillow
PyAutoGUI
pynput==1.7.8
pyperclip
